Cleaned up User routes
Also added Bootstrap CDN scripts.
