+++ views/index.handlebars
@@ -2,6 +2,7 @@
   <div class="container">
     <div class="row">
       <div class="col">
+        {{msg}}
         <h1>This is the Book Club App Home Page!</h1>
         <h2>Welcome to our app</h2>
         <p>Lorum Ipsum</p>
@@ -12,5 +13,5 @@
   </div>
 </div>
 
-<script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
-<script src="/js/index.js"></script>
+{{!-- <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script> --}}
+{{!-- <script src="/js/index.js"></script> --}}
