+++ views/layouts/main.handlebars
@@ -6,6 +6,15 @@
   <meta charset="utf-8">
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-beta.3/css/bootstrap.min.css">
   <link rel="stylesheet" href="/styles/styles.css" media="screen" title="no title" charset="utf-8">
+
+  <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"
+    integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1"
+    crossorigin="anonymous"></script>
+  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"
+    integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM"
+    crossorigin="anonymous"></script>
+
   <title>Home Page</title>
 </head>
 
@@ -27,6 +36,11 @@
                 <li class="nav-item">
                   <a class="nav-link" href="/">Home</a>
                 </li>
+
+                <li class="nav-item">
+                  <a class="nav-link" href="/users">Users</a>
+                </li>
+
                 <li class="nav-item">
                   <a class="nav-link" href="/clubs">Clubs</a>
                 </li>
@@ -61,14 +75,15 @@
       <div class="col">
         <div class="footer">
           <p>Footer content &copy; 2019<br>
-            <a href="https://www.freepik.com/free-photos-vectors/book">Book vector created by freepik -
-              www.freepik.com</a>
           </p>
         </div>
       </div>
     </div>
   </div>
-  <script src="/js/site.js"></script>
+  {{!-- <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
+    integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo"
+    crossorigin="anonymous"></script> --}}
+
 
 </body>
