+++ views/users.handlebars
@@ -1,0 +1,20 @@
+<div class="container">
+  <div class="row">
+    <div class="col-9">
+      <h1>Users</h1>
+      <div>
+      </div>
+      <div class="card mt-4">
+        <div class="card-body">
+          <h5 class="card-title">
+            <h3>All Users</h3>
+          </h5>
+          <div id="user-list"></div>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
+
+<script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
+<script src="/js/users.js"></script>
